Fix kitsu gif display
